+++ variables.tf
@@ -426,3 +426,12 @@ variable "enable_http2" {
   type        = bool
   default     = true
 }
+
+### Autoscaling
+
+variable "autoscaling" {
+  type        = list(map(string))
+  description = "Map containing autoscaling parameters"
+  default     = null
+}
+