+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+Fixed
+  * AZ-449: Update default value for `policy_name` parameter in `ssl_policy` block
+
 # v4.4.1 - 2021-02-23
 
 Fixed