+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Unreleased
 
 Added
-  * AZ-429 Add autoscaling
+  * AZ-429: Add autoscaling
 
 Fixed
   * AZ-449: Update default value for `policy_name` parameter in `ssl_policy` block